Extracted connection params
Modularised connection so that it can be referred to from multiple locations, this fixed the error thrown by sqlQueries.js
